assets/converter/requirements.txt
pdf2image
Pillow
pytesseract